Blood Bank + Records
This is the first version of this program it has some functionality but unlike my use of the word "functionality" this code would not be useful in the slightest. So to fix that some later updates for this code would be to add more functions for it. Such as actually making and storing records, adding the function to make and "record" newer user made usernames and passwords, keep live records of the "blood" in the "current facilities" (THESE ARE NOT REAL) and and a more modernized face for the project (Adding some Frontend coding for the User to see and better experience what is going on).