BloodBank+ Records
This code is not done just step one. The function of the current code is that it is a "very good" outline of a record keeping program. In the context of this code the function is to keep records and allow for 3 types of users to interact with this. Currently this version of the code does not do anything. It does not keep records nor does it update itself to have new information but hopefully this changes. 


Note: Although this is an outline there is a little bit of personal flare inside the code that should not be kept in further iterations of this code. Similar to many of the program that will be written by me the Outlines should be very fun to start , later as the program matures the sentiment of the program and programmer should as well.